# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "LIBRARY API documentation",
    "description": "Documentation of Library API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "contact@library-api.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "authorization",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/libraryapi/api/v1/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Get all members",
        "description": "This endpoint retrieves all the members in the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/members/{id}": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Get a member by ID",
        "description": "This endpoint retrieves a member based on the provided ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/addmember": {
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Add a new member",
        "description": "This endpoint adds a new member to the database",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phoneNumber": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/updatemember/{memberId}": {
      "put": {
        "tags": [
          "Members"
        ],
        "summary": "Update a member",
        "description": "This endpoint updates a member in the database based on the provided ID",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/deletemember/{memberId}": {
      "delete": {
        "tags": [
          "Members"
        ],
        "summary": "Deletes a member by ID",
        "description": "Deletes a member from the database by providing the member ID",
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/libraryapi/api/v1/services/getreservations": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "Get all reservations",
        "description": "Get all the reservations from the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/services/addreservation": {
      "post": {
        "tags": [
          "Reservations"
        ],
        "summary": "Add reservation",
        "description": "Add a reservation to the database",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "BookTitle": {
                  "example": "any"
                },
                "MemberName": {
                  "example": "any"
                },
                "ReservationDate": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/updatereservation/{reservationId}": {
      "put": {
        "tags": [
          "Reservations"
        ],
        "summary": "Update reservation",
        "description": "Update a reservation in the database",
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/deletereservation/{reservationId}": {
      "delete": {
        "tags": [
          "Reservations"
        ],
        "summary": "Delete reservation",
        "description": "Delete a reservation from the database",
        "parameters": [
          {
            "name": "reservationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
...
}